refactor: TableGroup, OrderTable 패키지 병합
