+++ src/main/java/kitchenpos/table/domain/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.tablegroup.repository;
+package kitchenpos.table.domain.repository;
 
-import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
