--- src/main/java/kitchenpos/tablegroup/exception/OrderTableNotFoundException.java
@@ -1,10 +1,0 @@
-package kitchenpos.tablegroup.exception;
-
-import kitchenpos.common.exception.KitchenposException;
-
-public class OrderTableNotFoundException extends KitchenposException {
-
-    public OrderTableNotFoundException(String message) {
-        super(message);
-    }
-}
