+++ src/test/java/kitchenpos/table/acceptance/OrderTableAcceptanceTest.java
@@ -31,9 +31,9 @@ import kitchenpos.product.ui.response.ProductResponse;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.ui.request.OrderTableGuestsRequest;
 import kitchenpos.table.ui.request.OrderTableRequest;
+import kitchenpos.table.ui.request.TableGroupRequest;
 import kitchenpos.table.ui.response.OrderTableResponse;
-import kitchenpos.tablegroup.ui.request.TableGroupRequest;
-import kitchenpos.tablegroup.ui.response.TableGroupResponse;
+import kitchenpos.table.ui.response.TableGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
