+++ src/test/java/kitchenpos/table/application/OrderTableServiceTest.java
@@ -12,17 +12,17 @@ import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.repository.OrderRepository;
 import kitchenpos.order.exception.InvalidOrderStatusException;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.domain.repository.OrderTableRepository;
+import kitchenpos.table.domain.repository.TableGroupRepository;
 import kitchenpos.table.exception.InvalidNumberOfGuestsException;
 import kitchenpos.table.exception.OrderTableEmptyException;
 import kitchenpos.table.exception.OrderTableNotEmptyException;
 import kitchenpos.table.exception.OrderTableNotFoundException;
-import kitchenpos.table.domain.repository.OrderTableRepository;
 import kitchenpos.table.ui.request.OrderTableEmptyRequest;
 import kitchenpos.table.ui.request.OrderTableGuestsRequest;
 import kitchenpos.table.ui.request.OrderTableRequest;
 import kitchenpos.table.ui.response.OrderTableResponse;
-import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.tablegroup.repository.TableGroupRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
